+++ src/types.jl
@@ -1,6 +1,8 @@
 export LaplacianOptModel
 
 """
+    LaplacianOptModel
+
 The composite mutable struct, `LaplacianOptModel`, holds dictionaries for input data, abstract JuMP model for optimization,
 variable references and result from solving the JuMP model.
 """
@@ -24,3 +26,27 @@ mutable struct LaplacianOptModel
 
 end
 
+"""
+    GraphData
+
+The composite mutable struct, `GraphData`, holds matrices of adjacency matrix, laplacian matrix,
+fiedler vector and algebraic connectivity.
+"""
+mutable struct GraphData
+    adjacency::Array{Float64}
+    laplacian::Array{Float64}
+    fiedler::Vector{Float64}
+    ac::Number
+    
+    function GraphData(adjacency::Array{Float64})
+        adj_matrix = adjacency
+        laplacian = LO.laplacian_matrix(adjacency)
+        fiedler = LO.fiedler_vector(adjacency)
+        ac = LO.algebraic_connectivity(adjacency)
+
+        graph_data = new(adj_matrix, laplacian, fiedler, ac)
+
+        return graph_data
+        
+    end
+end